lots o stuff
finally got around to getting Eclipse and GitHub connected. Since last
commit I did a lot and made many updates. And the fun part is, now you
can see them as they are on GitHub. I was planning on waiting for the
first pre-release before connecting to github but i got anxious so here
it is! Pre-pre-release code, yippee!!