I FINISHED THE SETUP PHASE!!
oh yeah go me! I did all the little things that needed to be done to
finish the game's setup phase, and now it's done. Next comes the main
loop and making new classes, but first I'm gonna move the setup to a new
method in a new class to make the main method easy to read and
understand.