unmixed up the everything!
i decided the whole 'Setup' class was garbage so I deleted it and moved
setup() back where it belongs, in 'Main'. I also did a million other
tiny things that made 'Main' a whole universe better than it was. I also
utilized Eclipse's collapsing feature to collapse my methods so I was
not bombarded with too many lines of code all the time. I should have
been doing that from the beggining. Also I got a new router which messed
up my internet so now I am stuck on WiFi for a bit instead of my regular
fast ethernet cord. At least the new router makes the wireless a whole
lot faster than it used to be. And that should be all for this commit.
Next step (for real): making the main loop and testing some frameworks
for the item/monster/etc data storage and access.