succesfully made it ask you a number
at first glance running the program, it just seems as if I have made it
ask you yet another question, except this time I made it way
overcomplicated and allowed to turn on and off certain options with it
still looking good to the user. the system should be a flexible way to
add and remove options each turn depending on stats. Someday, when I
implement an "extension" system for someone else to make a sort of mod,
They can code in their own options extremely easily. This is because
each option is an object that you can make more of and add to the list
of shown options every turn. more detail is coming in a blog post as i
realize this is a much much too long commit message. btw it is a long
term goal of mine to make commit messages that aren't too long or too
short or have too much or too little info in them. right now im at
ground zero on that goal. either way bye.