harder to win
now you have like a 10% chance of jumping the volcano or something...
idk random number generator is inclusive of 0 so i had to put in 99
instead of 100 to get a number between 0-99 rather than 1-100 like i
wanted. i decided if the number is greater than 79 you win, so thats
around 10-20% chance you do it successfully. enjoy this update!