this inventory stuff is getting to my head
I was going to take a break and work on something else,
but when I switched it to the "non-fancy" method,
it still gave me a very similar error to the one in the "fancy" method.
I am going to re-write these two methods, starting with the "non-fancy" one,
and hopefully learn from my mistakes. Then I will take a break
from all of this and forget about
inventory stuff forever
because it absolutely sucks.