changed player initialization, got rid of old inventory
You are no longer prompted for a starting level and balance, it is 0 by default.
I will probably add command line options to set this stuff in the future
The old inventory is completely gone, there should be no trace,
although I do need to go and check it over one more time.
Most all of the variables used to initialize the player have either been removed,
like the above money and level,
or moved to the Data class.
The startingInventory is now more modder-friendly,
with a static block you can use to add items to it. There is one example
commented out in there.
Finally, I made all these changes a while ago but never committed them so they didn't reach GitHub.
I hope you enjoy all the changes. Feel free to make an issue in GitHub to let me know
of a problem that needs to be fixed!
-foxler2010